+++ internal/bcd/forge/string.go
@@ -11,7 +11,7 @@ import (
 type String base.Node
 
 // Unforge -
-func (s *String) Unforge(data []byte) (int, error) {
+func (s *String) Unforge(data []byte) (uint32, error) {
 	l := new(length)
 	n, err := l.Unforge(data)
 	if err != nil {
@@ -20,7 +20,7 @@ func (s *String) Unforge(data []byte) (int, error) {
 
 	data = data[n:]
 
-	if len(data) < l.Value {
+	if uint32(len(data)) < l.Value { // #nosec G115 -- unforged data is bounded by protocol operation size limits, never close to uint32 max
 		return 4, errors.Wrap(ErrTooFewBytes, fmt.Sprintf("String.Unforge: %d < %d", len(data), l.Value))
 	}
 
@@ -34,7 +34,7 @@ func (s *String) Unforge(data []byte) (int, error) {
 func (s *String) Forge() ([]byte, error) {
 	l := new(length)
 	b := []byte(*s.StringValue)
-	l.Value = len(b)
+	l.Value = uint32(len(b)) // #nosec G115 -- forged string is bounded by protocol operation size limits, never close to uint32 max
 	data, err := l.Forge()
 	if err != nil {
 		return nil, err
